Move entry into sub module
